feat: provide `init` command
We begin by providing an `init` command, which may be subsumed into a `deploy` command; however, the
setup that is performed in the `init` function will still be necessary even if the command goes
away.

This is doing everything that's being done in the `Justfile` in the other testnet tool repository,
namely:
* Creating a new Terraform workspace
* Obtaining the RPC client (used for querying the genesis node to get its peer ID)
* Generating inventory for Ansible

Right now it only provides an implementation for Digital Ocean, not AWS.

Behaviour driven unit tests make use of traits/mocks to avoid calls to the Terraform process.

One other important note here is that the code has not be written with the intention of being cross
platform, since Ansible only runs on Linux/macOS.