ci: simple merge workflow
Provides all the basic functionality for a simple merging workflow: build, unit test, clippy and so
on.

Removed a utility script that was accidentally committed.

This also fixes a test setup issue. The way `assert_fs` works is that when the `TempDir` variable
reference is dropped, the directory it created gets cleaned up and deleted. It happened to be the
case that I had switched on `into_persistent` for debugging in my local setup, which was keeping the
directory alive, but once I switched this off, it started dropping the `TempDir` at the end of the
`setup_working_directory` function call. The fix was to keep the returned variable reference alive
until the end of the test, by adding a `drop` function call at the end.