feat: provide `deploy` command
The `init` command is subsumed as a new `deploy` command because there's no real reason to have
these separately; however, all the functionality from the `init` function in the `TestnetDeploy`
struct is still retained.

At this point, only the Terraform run is being performed. The Ansible run is still to follow.