feat: provide `clean` command
This command will run a `terraform destroy`, delete the Terraform workspace, and clear out the
generated Ansible inventory files.

It will later be extended for AWS to clear out the created keypair.

This commit also changes the library code in the crate to make use of the `Error` type defined in
the crate rather than using `eyre` for generic errors. In doing so, we introduced a few specific
errors for failures in the deployment process.