tests: reorganise unit tests
There were over 700 lines of tests and setup code inside a single `test` module, which was becoming
unwieldy to work were.

This splits the setup code into one module, then a module for each function on `TestnetDeploy` that
is under test. With the use of modules, the function name was removed from the test case names.