feat: ssh availability check
It's possible that before we run Ansible, the SSH service is not yet running on the genesis node.
This is something we seen happen occasionally, so we added in a little loop to sit and wait for it
to become available.

We only check this for the genesis run because the remaining nodes are provisioned after genesis, by
which point the SSH service should be available.