chore: respond to feedback
Addresses two issues that were noted during review:

* Provide more helpful error message when inventory list is empty
* Set other environment variables for Digital Ocean based on the `DO_PAT` variable

For the first issue, there are a few different ways in which it might be possible for the
environment to be non-existent. These were documented with comments in the code, which you can see
in the diff.

In addition, license headers were applied where they were missing.