chore: provide image definition for a build machine
Uses Packer to define a VM image which has a Rust environment pre-installed on it.

Since we already had a build setup defined in Ansible, we use Ansible as the provisioner for the
node. The way that Packer uses Ansible is quite strange. It runs it on the VM it spins up, rather
than running it on the machine where Packer is executing. So Ansible has to be installed first. This
was done using a script that uses a retry loop, since I was running into the usual problem with the
Apt package manager and file locks.