feat: deploy faucet as a service
The faucet is now introduced into the deployment process.

Since we don't release it as a binary, we need to build it. This means the build VM is no longer an
optional component. The build VM has been changed to use a droplet snapshot which already has a Rust
installation on it, which significantly reduces the overall deploy time.

The `build_node` role was changed to `build_safe_network_binary`, which can build any binary in the
workspace and upload it to S3. We use this to build `faucet`, then use it to build `safenode` if the
user has used the `--branch` and `--org` arguments.

For an unknown reason there seems to be some kind of issue with systemd and the journaling service
on the Ubuntu 22.10 base image. The `systemd-journald` service needs to be restarted before you see
any output in the journal, and the faucet service seems to need to be restarted outside of the
Ansible module before it will actually begin executing properly.