feat: use base image for nodes
There seems to be some intermittent problems at the moment with connectivity between DO machines and
the apt repositories. Here we introduce a node image that has the general `apt-get update` done in
advance, as well as having several packages installed.

It also installs logstash and the AWS plugin, which should significantly reduce deployment time.

Also increase the SSH timeout values from 5 to 30 seconds and set Ansible forks to 20. This was fine
to be set to 50 on Linux, but didn't seem to work very well on macOS if it was 30 or above. Trying
20. Might need to vary this value on the command line.