feat: provide log retrieval commands
Log retrieval has two components: obtaining the files from S3 and reassembling them. There is a
separate subcommand for each.

Obtaining the files is a recursive process that involves visiting the hierarchy of folders in the
bucket. Since there are lots of folders, this can be time consuming for a large testnet.

Logstash is forwarding the files to the bucket in parts, every 30 minutes. For this reason, they
need to be reassembled back into a single file. This is based on the part numbers in the file names.
The use of different GUIDs on each file part is done intentionally by Logstash.

The reason these have been implemented as separate subcommands is so that we can continually sync
the logs from S3 if that is necessary, without re-downloading the parts we've already retrieved. The
reassembly process removes the original part files.

There has also been some simple logging enabled in this commit, just using the `env_logger` crate,
since we don't really need anything sophisticated for this program.