refactor: download rpc client direct from S3
Rather than downloading the RPC client via HTTPs, it is retrieved directly from S3 using the AWS
SDK.

This was done because a problem showed up in CI with the `reqwests` library being rejected by
`cargo-deny`, due to a security vulnerability in the dependency chain. It turned out that I had
actually already used the AWS SDK on another branch anyway, for retrieving the logs from S3. So I
was able to do this refactor and remove the dependency on `reqwest`.