feat: provide log copy command
Use an Ansible playbook to copy logs from all the remote machines to the `logs` directory on the
local machine. There are a few redundant directories that get copied as part of the directory
structure, so there's a Python script that clears these up.

This also switches on trace logging for the services by using `SN_LOG=all`.