feat: provide logs rm command
Remove the logs from a previous testnet run by deleting the entire folder from S3.

When a new testnet is created we check to see if logs from a previous run with the same name already
exist. We won't proceed unless the user deletes those logs. We offer the choice to delete or
retrieve them before deletion.

If we did proceed, the logs from the two testnets would be intermingled with each other and this
would lead to a confusing situation.