feat: provide `logstash` commands
A `logstash deploy` command is provided.

Like the deployment of a testnet, it uses Terraform to create the VMs and Ansible to provision them.
The initial setup is to use one `s-4vcpu-8gb` VM, but we use a count in the Terraform configuration
so we could launch multiple VMs if we decided we wanted horizontal scaling, which Logstash and
Filebeat support. Ansible provisions one or more VMs with a replicated Logstash setup.

Filebeat authenticates with Logstash using self-signed SSL certificates. The Ansible playbook for
Logstash generates the certificate on the controller then copies the certificate and private key to
the Logstash hosts. Similarly, the genesis and node playbooks copy the public certificate to their
hosts, and it is referenced in the Filebeat configuration.

A client for interacting with the Digital Ocean REST API was introduced. This is used for obtaining
the hosts for Logstash, but I also intend to use it in a follow up PR to shut down the build machine
after it has done its initial work.

A `logstash clean` command is provided.

This works in an almost identical way to the testnet clean, so much so that the code was refactored
to be shared between the two.