feat: upload test data command
Provide a command to upload a set of standard test data, which is an archive with about 100mb of
jpeg images. The process is roughly:

* Run the `deploy` command
* The client binary is now built along with the node and faucet then uploaded to S3
* The rest of the testnet spins up as usual
* Retrieve some tokens from the faucet
* Run the `upload-test-data` command
* The test data archive is downloaded from S3 and extracted
* The `safe` client build is downloaded from S3
* We use the `files upload` command to upload each file in the archive

After all that, you get a list of the files and their addresses, with the potential to share those.
Though the same list will also be made available through the `safe` client.