chore: create node image with filebeat installed
The installation of Filebeat during the deploy process was causing too much grief. Almost 100% of
the time, it would fail because the unattended upgrades service was running in the background and
preventing access to the package manager. It was also very slow to run.

The `filebeat` role is split in two: one for installation and one for config. The installation role
is applied to the node image, and the config role is applied during deploy.