fix: return public ip for droplets
For some reason, the previous implementation for listing droplets didn't take into account the fact
that both public and private IPs can be returned. This was sometimes causing the private IP to be
selected for the Logstash server.

A new test case was added to explicitly select the public address.