feat: provide `smoke-test` command
This command performs a basic smoke test on the deployed network. We upload ten small files,
increasing in size from 1k to 10k, and the re-download them all and make sure we got the same files
back.

In the process of doing this, we are also testing that the faucet is working and we get a connection
to the network.

This test only takes something like 30 seconds, whereas uploading the full test data can take 30+
minutes.