feat: provide `notify` command
Sends a notification to Slack, which is populated by details from the deployment inventory for the
testnet. It's based on supplying an incoming webhook URL, which is a secret value.

Rather than send in a notification for every testnet that goes up, the user needs to explicitly run
the command to send it in. A lot of experimental testnets won't be of interest to everyone.