feat: deploy using version numbers for binaries
Optionally supply version numbers, which will deploy the binaries built during the release process.
We still need to build the faucet binary because that's not a released artifact at the moment.

When the version number arguments are used, we skip building the `safenode` and `safe` binaries,
which does shave a decent amount of time off the overall process.

The other significant difference is that we pull the version binaries from the Github release rather
than S3, so we had to introduce another mechanism for downloading those. At the moment the release
process does not include uploading the versioned binaries to S3. Only the 'latest' version goes up
there.