chore: remove encrypted file
This file provided a secret key that was being used for the purpose of querying the current EC2
instance. That information was then being used when we were using Open Telemetry. Since we aren't
using that any more, we don't need this file.

This file was the reason for providing the Ansible Vault password. It would be possible to take out
the need for the password, but if we ever using encrypted files at any point in the future, we'd
need to add it all back in again. For now you should just be able to provide any random value for
the password. It shouldn't actually be used in the process.