chore: upgrade build and node images
A new image was generated for the build machine with the newest version of Rust on it. The Rust
installation is not applied during the provisioning process. It's installed along with some other
prerequisites in an attempt to save time in the deploy process.

The version of Ubuntu that the build machine was using is actually no longer available, so I changed
it to a slightly earlier version. I tried a later one, but Ansible wouldn't install on it.

The node image was also regenerated here because I thought it also had a Rust installation on it,
but it actually doesn't. It doesn't really do any harm to just generate a new version anyway.