feat: idempotency for node deployment
The introduction of the node manager broke the idempotent nature of the deployment process.

Now, when the genesis node is being created, it will only be created if there's not already an
existing genesis node. With respect to the remaining nodes, they use a count value to determine
whether any more need to be added. If the count is satisfied, the task to add the nodes will be
skipped.

The commit also introduced an argument to supply a custom build of the node manager using a URL.
This was useful for testing out experimental features of the node manager.