chore: upgrade node image to ubuntu 23.10
In an attempt to resolve the problem we've been having with the Apt package manager failing to
upgrade packages due to a background automated update process that takes a lock on a file, the node
image is updated from Ubuntu 22.04 to 23.10. In a few different deploys I did not see any locking
issues.

The mechanism for installing Ansible for the image generation can't use the package manager, since a
PPA is not yet available for 23.10, so it had to be installed with pip. However, it turns out `pip`
really doesn't like to install any system-wide packages on 23.10, so the `--prefix` argument had to
be used to force that.

A new role was created that separates the creation of the `safe` user from the prerequisites role.
This means we can run the prerequisites role during the image generation, but we then don't need to
run it during the deploy provisions, since it actually eats up a bit of time to attempt to install
new versions of the packages. The `safe` user can't be created with the correct public SSH keys
during the image generation, so it has to be run during each deploy provision.