feat: use the node manager to deploy the faucet
Previously the faucet had been setup as a service 'manually', but if we want it to be part of the
upgrade process, it needs to be deployed by the node manager.

As part of this change, a new `node-manager` role is extracted from the `node` role so that the node
manager setup can be easily applied to the faucet playbook. Both the `node` and `faucet` roles
assume the existence of the node manager.

The `node_manager_upgrade` playbook was renamed to `upgrade_nodes`, to better reflect the purpose of
it.