feat: include the faucet in upgrades
We also need to be able to upgrade or downgrade the faucet as part of the whole process, so it is
now added to the setup.

As part of this, the `upgrade` command was extended with several new arguments that enable a
downgrade. This was really the easiest way of testing the process, since we could deploy with the
latest version and then downgrade to a previous version. Things seemed to work OK.