refactor: use builder for extra vars documents
The commit for using the node manager to deploy the faucet introduced a new type,
`ExtraVarsDocBuilder`. This was in attempt to clean up the code in the `DeployCmd` which assembled
the extra variables JSON document that gets passed to Ansible.

This refactor reduces some duplication and I think makes the code a little easier to read and work
with.