chore: alphabetically arrange cmds and args
In my opinion, when you have so many commands and arguments, the best way to keep them tidy is to
arrange them alphabetically. The `clap` crate that provides the CLI will indicate which of the
arguments are mandatory, and if the user needs more information on the mandatory arguments, they can
consult the list alphabetically.

Also just tidied the documentation on some of the arguments.