chore: use version/branch arguments differently
To make things easier to manage, I've changed argument usage with respect to versioning and building
binaries: you either use versions or you build all the binaries.

In the case of building the binaries, things are as they were before: you need to supply the branch
and repo owner at the same time.

With respect to versioning, each of the version arguments are optional, but if any get used,
versions will be used for all binaries. For the versions not supplied, the latest version will be
obtained using the releases repository.