feat: provide `--safenode-manager-version` arg
Allows a specific version of the node manager to be deployed, which is useful for testing the
upgrade process.

I also changed the `Versioned` variant of the `SnCodeBaseType` enum to use `semver::Version` rather
than strings, because I accidentally specified a version number in the wrong format, and didn't know
about it until Ansible tried to provision the node manager. Using the `Version` type forces the
argument to be validated at an early stage.