refactor: narrow binary option for deployment
We previously had an `SnCodebaseType` enum with three variants: `Main`, `Branch` and `Versioned`.
Now that we can use the release repository to retrieve the latest versions for binaries, I didn't
see the necessity for a third option; we simply use a version or we build from source. With that in
mind, I renamed the enum to `BinaryOption` and provided `BuildFromSource` and `Versioned` variants.

I don't think there should be any functional change here.