chore: remove `--safe-version` arg from `deploy`
We don't need to specify the client version when we are using a versioned deployment. It only gets
used if we are running automated smoke tests, which we don't do at the moment anyway. In any case, I
updated the `smoke-test` and `upload-test-data` commands to supply a `--safe-version` argument, in
case we do end up using these again in the future.

Also changed the use of the `Error` type in the `manage_test_data` module in favour of just using
`eyre`. It's debatable that this whole crate should be changed to do this and remove the `lib`
module.