feat: upload network contacts to s3
When a deployment completes, the network contacts will automatically be uploaded to S3. By default,
the file will have the same name as the environment, but an optional argument allows it to be
specified.

During the deployment, it's possible that some nodes don't start. These are indicated by "-" entries
in the peer list. The process ensures they will not be added to the contacts list.