feat: provide `--beta-encryption-key` arg on `deploy` cmd
A secret key can optionally be supplied, but otherwise, a default will be used. The default is moved
out of Ansible and into the `testnet-deploy` binary. This is being provided to support supplying a
key for the initial beta launch.

A Clippy warning related to a large enum variant was suppressed. There are quite a lot of arguments
for the tool, but it doesn't seem like it's something problematic.