feat: configure metrics collection on `deploy` cmd
Metrics gathering is implemented for testnet deployments. At the moment, this is not conditionally
controlled, so it would be applied to any deployment. Metrics may not be desirable for quick
experimental deployments, so we can put a control around this soon. Using Telegraf, the metrics are
forwarded to an Influx setup.

Since we now have metrics, the simple `resource-usage` role is no longer applied. However, we will
keep the role and may enable it again for simple deploys.

The following changes enable the setup:

* The node VM base image was updated with a Telegraf installation. It's better to install it here to
  avoid issues with locks on the `apt` package manager. These can often cause deployments to fail.
* Node RPC ports run from 13000 onwards rather than being randomly selected.
* Node metrics ports run from 14000 onwards rather than being randomly selected.
* The `safenode_rpc_client` binary is made available on every node VM.
* A `metrics` role was added and applied. It clones the `network-dashboard` repository and uses that
  for configuring telegraf.
* The Telegraf service is started and enabled for automatic start upon OS reboot.