refactor: `DeployCmd` -> `DeployOptions`
The `TestnetDeploy` struct is also renamed to `TestnetDeployer`, which is a better indicator that
this struct contains behaviour rather than just state.

The `DeployCmd` is renamed to `DeployOptions`, and the behaviour for executing the deployment is
moved back into `TestnetDeployer`. We still define the `deploy` function in a separate file, which
is worth it because there is quite a lot of code. We can take advantage of the fact that Rust allows
us to implement the behaviour of a struct across multiple files.

The reason for the refactor is because I always found it a bit odd that a large object like
`TestnetDeployer` was being passed into a command object. In most situations, commands are objects
that only have state, providing the parameters for some operation that is generally performed by the
command's corresponding handler. The command handler has the heavy dependencies that are used to
execute the operation. In our case, the command handler is `TestnetDeployer`.