feat: merge bootstrap and 'normal' nodes into one environment
Terraform now deploys two different types of nodes: bootstrap and 'normal', where the former are
a set of larger machines that are used for an initial bootstrap. Being deployed as two different
types means they can be scaled independently, and this can be controlled by two arguments on the
`deploy` command: `--bootstrap-node-vm-count` and `--node-vm-count`.

Merging the nodes into one environment also removes the need for the `--bootstrap-peer` argument. If
we want to extend our deployment with more machines of either type, we can run the `deploy` command
again with updated values for the node counts, and any peers we need will already be in the
inventory.

Two aspects of inventory management were also changed. First, inventory generation was updated to
retrieve the bootstrap node information. Secondly, we update the deploy process to try and fetch
inventory at the beginning of the deployment, as well as generating it at the end. On an initial
run, an empty inventory will be returned, but on deploys to add new machines, the retrieved
inventory will represent the current state. This inventory can then be used to disable certain
things we don't need to run on subsequent deploys, e.g., provisioning the faucet.

Tests performed:
✅ Initial `deploy --bootstrap-node-vm-count 10 --node-vm-count 10` uses 10 VMs for each
✅ Next `deploy --bootstrap-node-vm-count 15 --node-vm-count 10` adds 5 bootstrap nodes and leaves
   normal nodes at 10
✅ Next `deploy --bootstrap-node-vm-count 15 --node-vm-count 15` adds 5 nodes and leaves bootstrap
   nodes at 15
✅ Next `deploy --bootstrap-node-vm-count 10 --node-vm-count 15` removes 5 bootstrap VMs and leaves
   nodes at 15
✅ Next `deploy --bootstrap-node-vm-count 10 --node-vm-count 10` removes 5 node VMs and leaves
   bootstrap nodes at 10