feat: retry when generating inventory
It seems the inventory generation can randomly fail at times. A small retry loop is added, and if
the attempts don't succeed, the program will not return an error. It will advise the use of the
`inventory` command or workflow. It's more likely than not that there is some intermittent failure
connecting to machines rather than something actually being wrong with the inventory generation
process. Therefore, we don't need to throw out a whole deploy, especially if it was a large one. We
can try to get the inventory again with the print workflow.