feat: separate bootstrap and node vm list
The bootstrap and node VMs are separated into two lists. This is in anticipation of being able to
treat them differently from a scaling point of view. We want to be able to check how many of each
type there are.

Due to the network_commands module using the full VM list, we add a `vm_list` function which will
retain the whole list.