refactor: organise ansible with additional modules
The `ansible` module is now reorganised using two additional submodules, `extra_vars` and
`provisioning`.

The `extra_vars` module houses the code for building the extra vars documents; that code is
surprisingly large, so it's useful to have it sit in its own file.

The `provisioning` module has code related to the preparation involved for performing playbook runs,
which again, can be unexpectedly involved. Sometimes it is also providing wrappers around
`AnsibleRunner` that remove the need to work with that directly. This code used to sit in
`TestnetDeployer`, but it is being moved here in anticipation of it being shared between network
deployment and upscaling operations. A new type is introduced, `AnsibleProvisioner`. This does the
additional prep work for running a playbook to provision something, then uses `AnsibleRunner` to
run the playbook.