feat: provide `upscale` cmd
As the name implies, this command enables a network to increase in size. The following can be scaled
independently:

* The number of bootstrap node VMs
* The number of generic node VMs
* The number of node services running on bootstrap node VMs
* The number of node services running on generic node VMs
* The number of uploader VMs

As a safety feature, the command also prevents the network from being downsized, and really, this is
why it's a separate command from `deploy`. It would have been possible to run the `deploy` command
multiple times to achieve the same effect, but it wouldn't prevent downsizing by mistake.

Created a new network with `deploy --bootstrap-node-vm-count 10 --node-vm-count 10 --node-count 10`.
The default node count on the bootstrap VMs is 1.

✅ `upscale --desired-bootstrap-node-vm-count 15` increases bootstrap VMs from 10 to 15
✅ `upscale --desired-bootstrap-node-count 2` increases bootstrap nodes from 1 to 2
✅ `upscale --desired-node-vm-count 15` increases node VMs from 10 to 15
✅ `upscale --desired-node-count 15` increases nodes from 10 to 15
✅ `upscale --desired-uploader-vm-count 10` increases uploader VMs from 5 to 10
✅ `upscale --desired-node-count 10` prevents downsizing nodes from 15 to 10
✅ `upscale --desired-node-vm-count 10` prevents downsizing node VMs from 15 to 10
✅ `upscale --desired-bootstrap-node-vm-count 10` prevents downsizing bootstrap VMs from 15 to 10
✅ `upscale --desired-bootstrap-node-count 1` prevents downsizing bootstrap nodes from 2 to 1