feat: provide `--interval` arg for `upgrade` cmd
The `upgrade` command on the node manager already has an `--interval` argument, but it wasn't
controllable from `testnet-deploy`. The default value used by the node manager is only 200ms, which
has been observed to be problematic in some cases.

Also fix a little bug that caused `testnet_deploy` to be incorrectly initialised.