feat: `upgrade` cmd purposed for use in workflow
The `upgrade` command was originally written to run from a dev environment, which assumes the
existence of an initialised Terraform workspace. An upgrade doesn't require a Terraform run, but the
Terraform workspace was being used to check for the existence of the environment specified. This
check is now performed using the deployment inventory.

The inventory involves fetching the node manager's registry file from each machine, which is done
using an Ansible playbook. This file can change frequently, because the monitoring infrastructure is
calling the `status` command of the node manager. This can cause the playbook to fail with a
checksum mismatch, so it is disabled.

There are also extensions to the process. Bootstrap nodes are included, and failures are
anticipated, particularly from upgrading generic nodes. It seems to be difficult to get every node
to upgrade and restart correctly, so the workflow will not return an error in the case of any
failures. The log needs to be inspected to see any failures, and what subsequent action to take, if
any.