feat: provide `status` command
The command uses an existing playbook to pull all the node registry files and serialize them to a
`NodeRegistry`, which contains version and status information. This code was in the inventory
service, but it has now been moved to `AnsibleProvisioner` to share it between the inventory service
and this new command.

The code is also used at the end of the upgrade command to retrieve the status after an upgrade
attempt.