feat: more reliable ssh configuration
Increase the timeout for both Ansible and SSH, and try a larger value for `ControlPersist`. If
`ConnectTimeout` was lower than the Ansible `timeout` value, I think it would effectively override
the Ansible value. Also attempting to turn of `ControlMaster`.

Also, specify `-vvvvv` when `--ansible-verbose` is used so that we can get the full SSH specification.