feat: provide `--custom-inventory` arg for `upgrade` cmd
Use this argument to pass a list of particular VMs for the upgrade to run against. It can be used to
target VMs that were 'unreachable' in previous runs. It saves time by narrowing the scope of
machines to run against.

It works by looking up the IP addresses of the VMs named using the argument, then writing those IP
addresses out to a custom inventory file.