feat: vary machine sizes and counts for environment types
An `EnvironmentType` enum is introduced, with three variants: `Development`, `Staging`, and
`Production`. The intention is for the network being deployed to have a specification that resembles
either of these three possibilities. The variant determines the number of droplets, their size, and
image IDs, and the number of node services on each VM. These values are stored in `tfvars` files
that correspond to each of the variants.

The environment type selection is persisted in the deployment inventory by storing it in a file in a
new S3 bucket. This is used so that the correct tfvars file gets used during upscale operations.
Without this, there would be a risk that droplet sizes could change, which would cause droplets to
be recreated. The file is deleted when the deployment is cleaned. For the same reason,
`production.tfvars` file retains the previous droplet image IDs.

All of the count args on the `deploy` command are turned into options. These can still be used to
override the values in the tfvars files. With Terraform, variables provided by the `-var` argument
take precedence over values provided by `-var-file`. However, the droplet sizes and image IDs cannot
be varied.

There are now separate droplet images for staging and production, because DO prohibits the creation
of a droplet that is smaller in size than the image. So it's just easier to have separate images.
Right now, the development environment type will use the same droplet images as staging, but we
could potentially vary that if we need to. For these reasons, the default image IDs have been
removed. It doesn't make sense for there to be defaults; it depends on the environment type. The
same thing applies for droplet sizes. It may seem strange to declare the variables without defaults,
but Terraform requires that.