chore: introduce release process
The process here is copied from one of our other non-workspace repositories. It is slightly
different here in that we only need to produce a binary for Linux, but otherwise it is the same.

The testnet tool also needs access to the Terraform and Ansible definitions, so it doesn't work as a
binary in isolation. We can add a second step to somehow distribute these files along with it. For
now, a tag of the repository can be checked out along with the binary.