chore: try other ssh options
Increase the timeout values to 10 minutes and specify a fairly small keep-alive interval.

Even though Ansible reports 'UNREACHABLE', I have managed to confirm that a connection does actually
take place. The status of the node manager's registry was misleading me to believe no connection was
established. The node manager does not update the registry for every node that is upgraded; it only
updates it at the end of the whole operation. So, if it gets cut off, the registry makes it look
like no nodes have changed version. I was able to confirm that they did change version, by running
`safenode --version` for several of the individual binaries.

I am therefore trying a larger timeout, and trying to keep the connection alive by sending packets
every 10 seconds.

The plugin for showing the host that Ansible is currently running against has served its purpose, so
I'm now removing it.