fix: try to ensure environment type value
For an unknown reason, when it comes back from S3, the environment type file is sometimes empty, or
it has a newline character.

Here we try to get rid of the newline, and if the file is empty, fallback to using the name of the
environment to determine the type.