feat: take failure into account for node registry retrieval
Sometimes, particularly on the production environment, the retrieval of node registries can fail.
The suspicion is that it's because Telegraf is in the process of writing to the file, and therefore
the file is incomplete, but, it's not completely clear. It seems to fail too often for that to be
the case.

Now we take the failure into account and return a list of the VMs for which we couldn't retrieve
registries.