chore: redefine specs for new production environment
These new settings are based on analysis of the metrics we are collecting.

The specifications for generic nodes are retained. For bootstrap nodes, they are downsized to 4xCPU
and 8GB of memory. The estimated cost reduction here is from $6400 to $2400 per month. On both
types, the number of node services will remain the same.

The auditor count is reset to 1 for the new environment.