refactor: make some version numbers optional
To setup a command that will deploy a new set of nodes that are bootstrapped from an existing
deployment, the auditor, faucet and safe versions will be set to optional.

The bootstrapped deployment will just supply some extra nodes, so it won't have a genesis node, a
faucet, an auditor, or any uploaders.