feat: extend deployment inventory with deployment type
The inventory for a deployment now stores the type of the deployment, in addition to the environment
type.

The deployment type is either `New` or `Bootstrap`, where the latter is a deployment that bootstraps
from one of the peers in an existing deployment. In the bootstrap deployment, there will be no
genesis, bootstrap, uploader or auditor VMs.