chore: create nat gateway image for prod
Using the bootstrap image did not work for the production configuration because the prod bootstrap
image is larger than the size of the machine we are trying to create for the gateway.

Had to change the Ubuntu version because it seems that Digital Ocean have now pulled the version we
were using. Also had to add some extra retries into the Ansible installation script to get the image
to build properly.