feat: provide `--safe-version` arg for `upscale` cmd
Since we don't record the versioning information that was used with the initial deployment, if we
are going to upscale the uploaders, we need to supply the version we want to use for the `safe`
binary.

This is a simple solution for now. In the future we could record the binary information in the
environment information that is on S3.