feat: provide `--additional-volume-size` arg for `deploy` cmd
An optional argument, that when specified, will add an additional volume to the node VMs.

The node manager will then use that volume as data and logging storage space. On Digital Ocean, the
disk is automatically formatted and mounted at a location that corresponds to the volume label. This
is unlike AWS, where you need to format and mount the volume yourself.

The extra vars document processing is also extended to add a boolean type. Boolean variables should
not be enclosed in quotes if you want Ansible to interpret them as a boolean rather than a string.