feat: provide `--additional-volume-size` arg for `upscale` cmd
An optional argument, that when specified, will add an additional volume to new node VMs when an
existing deployment is upscaled.

There is a caveat here: if the original deployment used an additional volume, the new nodes from the
upscale must also use an additional volume. This is because the Terraform configuration synchronises
the node VMs with the additional volumes on a one-to-one basis. So if you try to upscale without
also adding volumes, Terraform will actually delete volumes for the existing VMs. The easiest thing
is to just keep the two synchronised.

For this reason, an additional field was added to the `EnvironmentDetails` struct to keep track of
whether the original deploy used additional volumes. At the same time, the storage of the
`EnvironmentDetails` struct on S3 was refactored to serialize it to JSON.

With that new information available, the command will error if the original deploy used additional
volumes and they were not specified for the upscale.