feat: provision a custom evm network
The `deploy` command has now been updated to provision an EVM-based network.

For the custom EVM network that will be used for dev/staging networks, the RPC URL, payment
addresses and the deploy secret key are retrieved from the Anvil instance, then passed to each of
the provisions. The provisions use the new EVM-related arguments on the node manager to have the
nodes spin up with the retrieved details.

Those EVM arguments require a new build of the node manager, so until there is a new release,
deploys will need to build binaries from source.

The provisions of the faucet and the auditor have now been removed. The uploaders provision is
commented out in this commit, to be re-enabled when the uploaders are adapted to EVM.