feat: change the uploaders to use `autonomi_cli`
The uploader script and service is changed to use the `autonomi_cli`. To maintain compatibility with
the monitoring setup, some things, like the `safe` user, have not yet been changed.

Fetching the EVM testnet data was also put in a retry loop. It seems to be possible for this file to
take a little bit of time to become available.